# tech.core/tech.core.service.identity/doc/swagger.yaml
swagger: '2.0'
info:
  version: '1.0-SNAPSHOT'
  title: TC Service Identity
  license:
    name: MIT
    url: http://github.com/gruntjs/grunt/blob/master/LICENSE-MIT
host: api.topcoder.com
basePath: /v3
schemes:
- http
- https
securityDefinitions:
  bearer:
    type: apiKey
    name: Authorization
    in: header
paths:
...
  /groups/{groupId}:
    get:
      description: Fetch a particular group
      tags:
      - groups
      operationId: GroupResource.getObject
      security:
        - bearer: []
      produces:
      - application/json
      parameters:
      - name: groupId
        in: path
        required: true
        type: string
        description: the id of group to fetch
      - name: fields
        in: query
        required: false
        type: string
        description: |
          fields=fieldName1,fieldName2,...,fieldN  - parameter for
          choosing which fields of group that will be included in response.

          + id
          + createdAt
          + createdBy
          + updatedAt
          + updatedBy
          + name
          + description
      responses:
        200:
          description: OK
          schema:
            type: object
            properties:
              id:
                type: string
              version:
                type: string
              result:
                type: object
                properties:
                  success:
                    type: boolean
                  status:
                    type: integer
                  metadata:
                    type: object
                  content:
                    $ref: "#/definitions/GroupResponse"
        400:
          description: Invalid input.
          schema:
            $ref: "#/definitions/ErrorModel"
        401:
          description: If user not authenticated.
          schema:
            $ref: "#/definitions/ErrorModel"
        403:
          description: If the group is private and the user is not a member of the group and or an admin.
          schema:
            $ref: "#/definitions/ErrorModel"
        404:
          description: If the group does not exist.
          schema:
            $ref: "#/definitions/ErrorModel"
        500:
          description: Internal server error.
          schema:
            $ref: "#/definitions/ErrorModel"
    delete:
      description: Delete a group
      tags:
      - groups
      operationId: GroupResource.deleteObject
      security:
        - bearer: []
      produces:
      - application/json
      parameters:
      - name: groupId
        in: path
        required: true
        type: string
        description: the id of group to delete
      responses:
        200:
          description: OK. Response will contains deleted group.
          schema:
            type: object
            properties:
              id:
                type: string
              version:
                type: string
              result:
                type: object
                properties:
                  success:
                    type: boolean
                  status:
                    type: integer
                  metadata:
                    type: object
                  content:
                    $ref: "#/definitions/GroupResponse"
        400:
          description: Invalid input.
          schema:
            $ref: "#/definitions/ErrorModel"
        401:
          description: If user not authenticated.
          schema:
            $ref: "#/definitions/ErrorModel"
        403:
          description: If user does not have admin role.
          schema:
            $ref: "#/definitions/ErrorModel"
        404:
          description: If the group does not exist.
          schema:
            $ref: "#/definitions/ErrorModel"
        500:
          description: Internal server error.
          schema:
            $ref: "#/definitions/ErrorModel"
    put:
      description: Update group
      tags:
      - groups
      operationId: GroupResource.updateObject
      security:
        - bearer: []
      produces:
      - application/json
      parameters:
      - name: groupId
        in: path
        required: true
        type: string
        description: the id of group to update
      - in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/NewGroupBodyParam'
      responses:
        200:
          description: OK. Response will contains updated group.
          schema:
            type: object
            properties:
              id:
                type: string
              version:
                type: string
              result:
                type: object
                properties:
                  success:
                    type: boolean
                  status:
                    type: integer
                  metadata:
                    type: object
                  content:
                    $ref: "#/definitions/GroupResponse"
        400:
          description: Invalid input.
          schema:
            $ref: "#/definitions/ErrorModel"
        401:
          description: If user not authenticated.
          schema:
            $ref: "#/definitions/ErrorModel"
        403:
          description: If user does not have admin role.
          schema:
            $ref: "#/definitions/ErrorModel"
        404:
          description: If the group does not exist.
          schema:
            $ref: "#/definitions/ErrorModel"
        500:
          description: Internal server error.
          schema:
            $ref: "#/definitions/ErrorModel"
...
  /groups:
    get:
      description: Fetch groups by user or group. Omit the query parameters to fetch all groups.
      tags:
      - groups
      operationId: GroupResource.getObjects
      security:
        - bearer: []
      produces:
      - application/json
      parameters:
      - name: memberId
        in: query
        required: false
        type: string
        description: id of membership
      - name: membershipType
        in: query
        required: false
        type: string
        description: membership type. Use 'group' if memberId is an id of a group, otherwise 'user'
      responses:
        200:
          description: OK. Response will contain array of groups.
          schema:
            type: object
            properties:
              id:
                type: string
              version:
                type: string
              result:
                type: object
                properties:
                  success:
                    type: boolean
                  status:
                    type: integer
                  metadata:
                    type: object
                  content:
                    type: array
                    items:
                      $ref: "#/definitions/GroupResponse"
        400:
          description: Invalid input.
          schema:
            $ref: "#/definitions/ErrorModel"
        401:
          description: If user not authenticated.
          schema:
            $ref: "#/definitions/ErrorModel"
        403:
          description: If user does not have admin role.
          schema:
            $ref: "#/definitions/ErrorModel"
        404:
          description: If the group does not exist.
          schema:
            $ref: "#/definitions/ErrorModel"
        500:
          description: Internal server error.
          schema:
            $ref: "#/definitions/ErrorModel"
    post:
      description: Creation of new group
      tags:
      - groups
      operationId: GroupResource.createObject
      security:
        - bearer: []
      produces:
      - application/json
      parameters:
      - in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/NewGroupBodyParam'
      responses:
        200:
          description: OK. Response will contains newly created group.
          schema:
            type: object
            properties:
              id:
                type: string
              version:
                type: string
              result:
                type: object
                properties:
                  success:
                    type: boolean
                  status:
                    type: integer
                  metadata:
                    type: object
                  content:
                    $ref: "#/definitions/GroupResponse"
        400:
          description: Invalid input.
          schema:
            $ref: "#/definitions/ErrorModel"
        401:
          description: If user not authenticated.
          schema:
            $ref: "#/definitions/ErrorModel"
        403:
          description: If user does not have admin role.
          schema:
            $ref: "#/definitions/ErrorModel"
        500:
          description: Internal server error.
          schema:
            $ref: "#/definitions/ErrorModel"
...
  /users/{resourceId}:
    get:
      description: Get user by id
      tags:
      - users
      operationId: UserResource.getObject
      security:
        - bearer: []
      produces:
      - application/json
      parameters:
      - name: resourceId
        in: path
        required: true
        type: string
        description: user id
      - name: fields
        in: query
        required: false
        type: string
        description: |
          fields=fieldName1,fieldName2,...,fieldN  - parameter for
          choosing which fields of user that will be included in response.

          + id
          + createdAt
          + createdBy
          + updatedAt
          + updatedBy
          + handle
          + email
          + firstName
          + lastName
          + regSource
          + utmSource
          + utmMedium
          + utmCampaign
          + ssoLogin
          + profiles
      responses:
        200:
          description: OK
          schema:
            type: object
            properties:
              id:
                type: string
              version:
                type: string
              result:
                type: object
                properties:
                  success:
                    type: boolean
                  status:
                    type: integer
                  metadata:
                    type: object
                  content:
                    $ref: "#/definitions/UserResponse"
        400:
          description: Invalid input.
          schema:
            $ref: "#/definitions/ErrorModel"
        401:
          description: If user not authenticated.
          schema:
            $ref: "#/definitions/ErrorModel"
        404:
          description: If user does not exist.
          schema:
            $ref: "#/definitions/ErrorModel"
        500:
          description: Internal server error.
          schema:
            $ref: "#/definitions/ErrorModel"
    delete:
      description: |
        Delete user by id.

        It is not yet implemented (501).
      tags:
      - users
      operationId: UserResource.deleteObject
      security:
        - bearer: []
      produces:
      - application/json
      parameters:
      - name: resourceId
        in: path
        required: true
        type: string
        description: user id
      responses:
        501:
          description: Not yet implemented
    patch:
      description: Update user by id
      tags:
      - users
      operationId: UserResource.updateObject
      security:
        - bearer: []
      produces:
      - application/json
      parameters:
      - name: resourceId
        in: path
        required: true
        type: string
        description: user id
      - in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/NewUserBodyParam'
      responses:
        200:
          description: OK. Response will contain updated user.
          schema:
            type: object
            properties:
              id:
                type: string
              version:
                type: string
              result:
                type: object
                properties:
                  success:
                    type: boolean
                  status:
                    type: integer
                  metadata:
                    type: object
                  content:
                    $ref: "#/definitions/UserResponse"
        400:
          description: Invalid input.
          schema:
            $ref: "#/definitions/ErrorModel"
        401:
          description: If user not authenticated.
          schema:
            $ref: "#/definitions/ErrorModel"
        403:
          description: If user to perform this operation does not had admin role.
          schema:
            $ref: "#/definitions/ErrorModel"
        404:
          description: If user does not exist.
          schema:
            $ref: "#/definitions/ErrorModel"
        500:
          description: Internal server error.
          schema:
            $ref: "#/definitions/ErrorModel"
...
  /users:
    get:
      description: Fetch list of users
      tags:
      - users
      operationId: UserResource.getObjects
      security:
        - bearer: []
      produces:
      - application/json
      parameters: []
      responses:
        200:
          description: OK
          schema:
            type: object
            properties:
              id:
                type: string
              version:
                type: string
              result:
                type: object
                properties:
                  success:
                    type: boolean
                  status:
                    type: integer
                  metadata:
                    type: object
                  content:
                    type: array
                    items:
                      $ref: "#/definitions/UserResponse"
        400:
          description: Invalid input.
          schema:
            $ref: "#/definitions/ErrorModel"
        401:
          description: If user not authenticated.
          schema:
            $ref: "#/definitions/ErrorModel"
        403:
          description: If user to perform this operation does not had admin role.
          schema:
            $ref: "#/definitions/ErrorModel"
        500:
          description: Internal server error.
          schema:
            $ref: "#/definitions/ErrorModel"
    post:
      description: Create a user
      tags:
      - users
      operationId: UserResource.createObject
      produces:
      - application/json
      parameters:
      - in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/NewUserBodyParam'
      responses:
        200:
          description: OK. Response will contain newly created user.
          schema:
            type: object
            properties:
              id:
                type: string
              version:
                type: string
              result:
                type: object
                properties:
                  success:
                    type: boolean
                  status:
                    type: integer
                  metadata:
                    type: object
                  content:
                    type: array
                    items:
                      $ref: "#/definitions/UserResponse"
        400:
          description: Invalid input.
          schema:
            $ref: "#/definitions/ErrorModel"
        500:
          description: Internal server error.
          schema:
            $ref: "#/definitions/ErrorModel"
...
  /authorizations/{targetId}:
    get:
      description: Get an authorization by target id.
      tags:
      - authorizations
      operationId: AuthorizationResource.getObject
      produces:
      - application/json
      parameters:
      - name: targetId
        in: path
        required: true
        type: string
        description: target id
      - name: fields
        in: query
        required: false
        type: string
        description: |
          fields=fieldName1,fieldName2,...,fieldN  - parameter for
          choosing which fields of authorization that will be included in response.

          + id
          + createdAt
          + createdBy
          + updatedAt
          + updatedBy
          + handle
          + email
          + firstName
          + lastName
          + regSource
          + utmSource
          + utmMedium
          + utmCampaign
      responses:
        200:
          description: OK
          schema:
            type: object
            properties:
              id:
                type: string
              version:
                type: string
              result:
                type: object
                properties:
                  success:
                    type: boolean
                  status:
                    type: integer
                  metadata:
                    type: object
                  content:
                    $ref: "#/definitions/AuthorizationResponse"
        400:
          description: Invalid input.
          schema:
            $ref: "#/definitions/ErrorModel"
        401:
          description: If not authorized to perform this operation.
          schema:
            $ref: "#/definitions/ErrorModel"
        500:
          description: Internal server error.
          schema:
            $ref: "#/definitions/ErrorModel"
    put:
      description: |
        Get an authorization by target id.
        It is not yet implemented (501).
      tags:
      - authorizations
      operationId: AuthorizationResource.updateObject
      security:
        - bearer: []
      produces:
      - application/json
      parameters:
      - name: targetId
        in: path
        required: true
        type: string
        description: target id
      responses:
        501:
          description: Not yet implemented
          schema:
            type: object
    delete:
      description: Delete an authorization by target id
      tags:
      - authorizations
      operationId: AuthorizationResource.deleteObject
      security:
        - bearer: []
      produces:
      - application/json
      parameters:
      - name: targetId
        in: path
        required: true
        type: string
        description: target id
      responses:
        200:
          description: Represents a successful deletion, the response content will just be null
          schema:
            $ref: "#/definitions/NormalResponse"
        400:
          description: Invalid input.
          schema:
            $ref: "#/definitions/ErrorModel"
        401:
          description: If not authorized to perform this operation.
          schema:
            $ref: "#/definitions/ErrorModel"
        500:
          description: Internal server error.
          schema:
            $ref: "#/definitions/ErrorModel"
  /authorizations:
    post:
      description: Create an authorization
      tags:
      - authorizations
      operationId: AuthorizationResource.createObject
      produces:
      - application/json
      parameters:
      - in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/NewAuthorizationBodyParam'
      responses:
        200:
          description: OK
          schema:
            type: object
            properties:
              id:
                type: string
              version:
                type: string
              result:
                type: object
                properties:
                  success:
                    type: boolean
                  status:
                    type: integer
                  metadata:
                    type: object
                  content:
                    $ref: "#/definitions/AuthorizationResponse"
        400:
          description: Invalid input.
          schema:
            $ref: "#/definitions/ErrorModel"
        401:
          description: If not authorized to perform this operation.
          schema:
            $ref: "#/definitions/ErrorModel"
        403:
          description: If account is inactive.
          schema:
            $ref: "#/definitions/ErrorModel"
        404:
          description: If account is not found.
          schema:
            $ref: "#/definitions/ErrorModel"
        500:
          description: Internal server error.
          schema:
            $ref: "#/definitions/ErrorModel"
    delete:
      description: Delete authorization with target id of "1"
      tags:
      - authorizations
      operationId: AuthorizationResource.deleteobject
      security:
        - bearer: []
      produces:
      - application/json
      parameters: []
      responses:
        200:
          description: Represents a successful deletion, the response content will just be null
          schema:
            $ref: "#/definitions/NormalResponse"
        400:
          description: Invalid input.
          schema:
            $ref: "#/definitions/ErrorModel"
        401:
          description: If not authorized to perform this operation.
          schema:
            $ref: "#/definitions/ErrorModel"
        500:
          description: Internal server error.
          schema:
            $ref: "#/definitions/ErrorModel"
  /roles/{roleId}:
    get:
      description: Fetch a particular role information
      tags:
      - roles
      operationId: RoleResource.getObject
      security:
        - bearer: []
      produces:
      - application/json
      parameters:
      - name: roleId
        in: path
        required: true
        type: string
        description: role id
      - name: fields
        in: query
        required: false
        type: string
        description: |
          fields=fieldName1,fieldName2,...,fieldN  - parameter for
          choosing which fields of role that will be included in response.

          + id
          + createdAt
          + createdBy
          + updatedAt
          + updatedBy
          + roleName
          + subjects
      responses:
        200:
          description: OK
          schema:
            type: object
            properties:
              id:
                type: string
              version:
                type: string
              result:
                type: object
                properties:
                  success:
                    type: boolean
                  status:
                    type: integer
                  metadata:
                    type: object
                  content:
                    $ref: "#/definitions/RoleResponse"
        400:
          description: Invalid input.
          schema:
            $ref: "#/definitions/ErrorModel"
        401:
          description: If user not authenticated.
          schema:
            $ref: "#/definitions/ErrorModel"
        404:
          description: If role does not exist.
          schema:
            $ref: "#/definitions/ErrorModel"
        500:
          description: Internal server error.
          schema:
            $ref: "#/definitions/ErrorModel"
    put:
      description: Update a particular role
      tags:
      - roles
      operationId: RoleResource.updateObject
      security:
        - bearer: []
      produces:
      - application/json
      parameters:
      - name: roleId
        in: path
        required: true
        type: string
        description: role id
      - in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/NewRoleBodyParam'
      responses:
        200:
          description: OK
          schema:
            type: object
            properties:
              id:
                type: string
              version:
                type: string
              result:
                type: object
                properties:
                  success:
                    type: boolean
                  status:
                    type: integer
                  metadata:
                    type: object
                  content:
                    $ref: "#/definitions/RoleResponse"
        400:
          description: Invalid input.
          schema:
            $ref: "#/definitions/ErrorModel"
        401:
          description: If user not authenticated.
          schema:
            $ref: "#/definitions/ErrorModel"
        403:
          description: If user does not have admin role.
          schema:
            $ref: "#/definitions/ErrorModel"
        404:
          description: If role does not exist.
          schema:
            $ref: "#/definitions/ErrorModel"
        500:
          description: Internal server error.
          schema:
            $ref: "#/definitions/ErrorModel"
    delete:
      description: Delete a role
      tags:
      - roles
      operationId: RoleResource.deleteObject
      security:
        - bearer: []
      produces:
      - application/json
      parameters:
      - name: roleId
        in: path
        required: true
        type: string
        description: role id
      responses:
        200:
          description: OK. Response will contain id of deleted role.
          schema:
            type: object
            properties:
              id:
                type: string
              version:
                type: string
              result:
                type: object
                properties:
                  success:
                    type: boolean
                  status:
                    type: integer
                  metadata:
                    type: object
                  content:
                    type: object
                    properties:
                      id:
                        type: string
        400:
          description: Invalid input.
          schema:
            $ref: "#/definitions/ErrorModel"
        401:
          description: If user not authenticated or does not have admin role.
          schema:
            $ref: "#/definitions/ErrorModel"
        403:
          description: If user does not have admin role.
          schema:
            $ref: "#/definitions/ErrorModel"
        404:
          description: If role does not exist.
          schema:
            $ref: "#/definitions/ErrorModel"
        500:
          description: Internal server error.
          schema:
            $ref: "#/definitions/ErrorModel"
...
definitions:
...
  NewUserBodyParam:
    type: object
    properties:
      param:
        $ref: "#/definitions/User"
...
  NewGroupBodyParam:
    type: object
    properties:
      param:
        $ref: "#/definitions/Group"
...
  UserResponse:
    type: object
    properties:
      id:
        type: integer
      createdAt:
        type: string
        format: date-time
      createdBy:
        type: string
      updatedAt:
        type: string
        format: date-time
      updatedBy:
        type: string
      handle:
        type: string
      email:
        type: string
      firstName:
        type: string
      lastName:
        type: string
      ssoLogin:
        type: boolean
      credential:
        $ref: "#/definitions/Credential"
      profiles:
        type: array
        items:
          $ref: "#/definitions/UserProfile"
      status:
        type: string
      country:
        $ref: "#/definitions/Country"
      regSource:
        type: string
      utmSource:
        type: string
      utmMedium:
        type: string
      utmCampaign:
       type: string
...
  GroupResponse:
    type: object
    properties:
      id:
        type: integer
      createdAt:
        type: string
        format: date-time
      createdBy:
        type: string
        format: integer
      updatedAt:
        type: string
        format: date-time
      updatedBy:
        type: string
      name:
        type: string
      description:
        type: string
      privateGroup:
        type: boolean
      selfRegister:
        type: boolean
      subGroups:
        type: array
        items:
            $ref: '#/definitions/GroupResponse' 
      parentGroup:
        $ref: '#/definitions/GroupResponse' 
...
  AuthorizationResponse:
    type: object
    properties:
      id:
        type: integer
      createdAt:
        type: string
        format: date-time
      createdBy:
        type: string
        format: integer
      updatedAt:
        type: string
        format: date-time
      updatedBy:
        type: string
      token:
        type: string
      refreshToken:
        type: string
      target:
        type: string
      externalToken:
        type: string
      zendeskJwt:
        type: string
...
  ErrorModel:
    type: object
    properties:
      id:
        type: string
        description: unique id identifying the request
      version:
        type: string
      result:
        type: object
        properties:
          success:
            type: boolean
          status:
            description: http status code
            type: number
          metadata:
            $ref: '#/definitions/ResponseMetadata'
          content:
            type: object
  NormalResponse:
    title: Response with null or a single string
    type: object
    properties:
      id:
        type: string
        description: unique id identifying the request
      version:
        type: string
      result:
        type: object
        properties:
          success:
            type: boolean
          status:
            type: number
            description: http status code
          metadata:
            $ref: "#/definitions/ResponseMetadata"
          content:
            type: object
            description: a string or null
...